move EnergyPerM2PctCommune to common energy
