+++ src/climatevision/generator/business2018/energy_demand/__init__.py
@@ -3,12 +3,9 @@
 from dataclasses import dataclass
 
 from ...inputs import Inputs
-from ...common.energy import Energy
+from ...common.energy import Energy, EnergyPerM2PctCommune
 
-from .dataclasses import (
-    EnergyPerM2WithBuildings,
-    EnergyPerM2PctCommune,
-)
+from .dataclasses import EnergyPerM2WithBuildings
 
 
 @dataclass(kw_only=True)
