+++ src/climatevision/generator/common/energy.py
@@ -34,3 +34,16 @@ class EnergyPerM2(Energy):
         self,
     ):
         self.ratio_energy_to_m2 = div(self.energy, self.area_m2)
+
+
+@dataclass(kw_only=True)
+class EnergyPerM2PctCommune(EnergyPerM2):
+    pct_x: float
+
+    total: InitVar[EnergyPerM2]
+
+    def __post_init__(self, total: EnergyPerM2):  # type: ignore
+        self.area_m2 = total.area_m2 * self.pct_x
+        self.energy = total.energy * self.pct_x
+
+        EnergyPerM2.__post_init__(self)
